Added automated commentators
Unfortunately, naive ui input component doesn't really support v-model all that well so we're left with using the default html input field which doesn't make it look nice. However, it will still update in real time.
